Implemented alternative input methods
Still some work to do; these changes are untested as of yet (though they
compile). Part of the purpose of this refactoring was to enable
find_clouds_and_align to be multithreaded, and that still needs to be
completed. Also, giving the option to take interleaved FASTQs from
stdin, or gzip'd FASTQs in any format, will be somewhat necessary.