Refactored to enable multithreading
All code within find_clouds_and_align should now be threadsafe, with
lock-protected access to any global variables and input/output files.
However, implementing multithreading might be more difficult. The
parallel pipe operator ||> might be of use, but it doesn't appear to be
configurable to the number of threads, at least not through an easy Seq
interface. We might be able to use _C.omp_set_num_threads instead,
though. Then the only refactoring that needs to be done is parallelizing
the main loop in find_clouds_and_align, which should be relatively
straightforward.